Move some classes into separate files, in particular simple Match related ones and ExpressionTester
