+++ Source/DafnyCore/AST/Expressions/Expressions.cs
@@ -2334,71 +2334,6 @@ public class ITEExpr : Expression {
   }
 }
 
-public class MatchExpr : Expression {  // a MatchExpr is an "extended expression" and is only allowed in certain places
-  private Expression source;
-  private List<MatchCaseExpr> cases;
-  public readonly MatchingContext Context;
-  [FilledInDuringResolution] public readonly List<DatatypeCtor> MissingCases = new List<DatatypeCtor>();
-  public readonly bool UsesOptionalBraces;
-  public MatchExpr OrigUnresolved;  // the resolver makes this clone of the MatchExpr before it starts desugaring it
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Source != null);
-    Contract.Invariant(cce.NonNullElements(Cases));
-    Contract.Invariant(cce.NonNullElements(MissingCases));
-  }
-
-  public MatchExpr(IToken tok, Expression source, [Captured] List<MatchCaseExpr> cases, bool usesOptionalBraces, MatchingContext context = null)
-    : base(tok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(source != null);
-    Contract.Requires(cce.NonNullElements(cases));
-    this.source = source;
-    this.cases = cases;
-    this.UsesOptionalBraces = usesOptionalBraces;
-    this.Context = context is null ? new HoleCtx() : context;
-  }
-
-  public Expression Source {
-    get { return source; }
-  }
-
-  public List<MatchCaseExpr> Cases {
-    get { return cases; }
-  }
-
-  // should only be used in desugar in resolve to change the source and cases of the matchexpr
-  public void UpdateSource(Expression source) {
-    this.source = source;
-  }
-
-  public void UpdateCases(List<MatchCaseExpr> cases) {
-    this.cases = cases;
-  }
-
-  public override IEnumerable<INode> Children => new[] { source }.Concat<INode>(cases);
-
-  public override IEnumerable<Expression> SubExpressions {
-    get {
-      yield return Source;
-      foreach (var mc in cases) {
-        yield return mc.Body;
-      }
-    }
-  }
-
-  public override IEnumerable<Type> ComponentTypes {
-    get {
-      foreach (var mc in cases) {
-        foreach (var bv in mc.Arguments) {
-          yield return bv.Type;
-        }
-      }
-    }
-  }
-}
-
 /// <summary>
 /// A CasePattern is either a BoundVar or a datatype constructor with optional arguments.
 /// Lexically, the CasePattern starts with an identifier.  If it continues with an open paren (as
@@ -2471,203 +2406,6 @@ public class CasePattern<VT> where VT : IVariable {
         }
       }
     }
-  }
-}
-
-public abstract class MatchCase : IHasUsages {
-  public readonly IToken tok;
-  [FilledInDuringResolution] public DatatypeCtor Ctor;
-  public List<BoundVar> Arguments; // created by the resolver.
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(tok != null);
-    Contract.Invariant(Ctor != null);
-    Contract.Invariant(cce.NonNullElements(Arguments));
-  }
-
-  public MatchCase(IToken tok, DatatypeCtor ctor, [Captured] List<BoundVar> arguments) {
-    Contract.Requires(tok != null);
-    Contract.Requires(ctor != null);
-    Contract.Requires(cce.NonNullElements(arguments));
-    this.tok = tok;
-    this.Ctor = ctor;
-    this.Arguments = arguments;
-  }
-
-  public IToken NameToken => tok;
-  public abstract IEnumerable<INode> Children { get; }
-  public IEnumerable<IDeclarationOrUsage> GetResolvedDeclarations() {
-    return new[] { Ctor };
-  }
-}
-
-public class MatchCaseExpr : MatchCase {
-  private Expression body;
-  public Attributes Attributes;
-  public readonly bool FromBoundVar;
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(body != null);
-  }
-
-  public override IEnumerable<INode> Children => Arguments.Concat<INode>(new[] { body });
-
-  public MatchCaseExpr(IToken tok, DatatypeCtor ctor, bool FromBoundVar, [Captured] List<BoundVar> arguments, Expression body, Attributes attrs = null)
-    : base(tok, ctor, arguments) {
-    Contract.Requires(tok != null);
-    Contract.Requires(ctor != null);
-    Contract.Requires(cce.NonNullElements(arguments));
-    Contract.Requires(body != null);
-    this.body = body;
-    this.Attributes = attrs;
-    this.FromBoundVar = FromBoundVar;
-  }
-
-  public Expression Body {
-    get { return body; }
-  }
-
-  // should only be called by resolve to reset the body of the MatchCaseExpr
-  public void UpdateBody(Expression body) {
-    this.body = body;
-  }
-}
-/*
-MatchingContext represents the context
-in which a pattern-match takes place during pattern-matching compilation
-
-MatchingContext is either:
-1 - a HoleCtx
-    standing for one of the current selectors in pattern-matching compilation
-2 - A ForallCtx
-    standing for a pattern-match over any expression
-3 - an IdCtx of a string and a list of MatchingContext
-    standing for a pattern-match over a constructor
-4 - a LitCtx
-    standing for a pattern-match over a constant
-*/
-public abstract class MatchingContext {
-  public virtual MatchingContext AbstractAllHoles() {
-    return this;
-  }
-
-  public MatchingContext AbstractHole() {
-    return this.FillHole(new ForallCtx());
-  }
-
-  public virtual MatchingContext FillHole(MatchingContext curr) {
-    return this;
-  }
-}
-
-public class LitCtx : MatchingContext {
-  public readonly LiteralExpr Lit;
-
-  public LitCtx(LiteralExpr lit) {
-    Contract.Requires(lit != null);
-    this.Lit = lit;
-  }
-
-  public override string ToString() {
-    return Printer.ExprToString(Lit);
-  }
-}
-
-public class HoleCtx : MatchingContext {
-  public HoleCtx() { }
-
-  public override string ToString() {
-    return "*";
-  }
-
-  public override MatchingContext AbstractAllHoles() {
-    return new ForallCtx();
-  }
-
-  public override MatchingContext FillHole(MatchingContext curr) {
-    return curr;
-  }
-}
-
-public class ForallCtx : MatchingContext {
-  public ForallCtx() { }
-
-  public override string ToString() {
-    return "_";
-  }
-}
-
-public class IdCtx : MatchingContext {
-  public readonly String Id;
-  public readonly List<MatchingContext> Arguments;
-
-  public IdCtx(String id, List<MatchingContext> arguments) {
-    Contract.Requires(id != null);
-    Contract.Requires(arguments != null); // Arguments can be empty, but shouldn't be null
-    this.Id = id;
-    this.Arguments = arguments;
-  }
-
-  public IdCtx(KeyValuePair<string, DatatypeCtor> ctor) {
-    List<MatchingContext> arguments = Enumerable.Repeat((MatchingContext)new HoleCtx(), ctor.Value.Formals.Count).ToList();
-    this.Id = ctor.Key;
-    this.Arguments = arguments;
-  }
-
-  public override string ToString() {
-    if (Arguments.Count == 0) {
-      return Id;
-    } else {
-      List<string> cps = Arguments.ConvertAll<string>(x => x.ToString());
-      return string.Format("{0}({1})", Id, String.Join(", ", cps));
-    }
-  }
-
-  public override MatchingContext AbstractAllHoles() {
-    return new IdCtx(this.Id, this.Arguments.ConvertAll<MatchingContext>(x => x.AbstractAllHoles()));
-  }
-
-  // Find the first (leftmost) occurrence of HoleCtx and replace it with curr
-  // Returns false if no HoleCtx is found
-  private bool ReplaceLeftmost(MatchingContext curr, out MatchingContext newcontext) {
-    var newArguments = new List<MatchingContext>();
-    bool foundHole = false;
-    int currArgIndex = 0;
-
-    while (!foundHole && currArgIndex < this.Arguments.Count) {
-      var arg = this.Arguments.ElementAt(currArgIndex);
-      switch (arg) {
-        case HoleCtx _:
-          foundHole = true;
-          newArguments.Add(curr);
-          break;
-        case IdCtx argId:
-          MatchingContext newarg;
-          foundHole = argId.ReplaceLeftmost(curr, out newarg);
-          newArguments.Add(newarg);
-          break;
-        default:
-          newArguments.Add(arg);
-          break;
-      }
-      currArgIndex++;
-    }
-
-    if (foundHole) {
-      while (currArgIndex < this.Arguments.Count) {
-        newArguments.Add(this.Arguments.ElementAt(currArgIndex));
-        currArgIndex++;
-      }
-    }
-
-    newcontext = new IdCtx(this.Id, newArguments);
-    return foundHole;
-  }
-
-  public override MatchingContext FillHole(MatchingContext curr) {
-    MatchingContext newcontext;
-    ReplaceLeftmost(curr, out newcontext);
-    return newcontext;
   }
 }
