+++ Source/DafnyCore/AST/Statements/Statements.cs
@@ -884,126 +884,6 @@ public class RefinedWhileStmt : WhileStmt {
   }
 }
 
-public class MatchStmt : Statement {
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(Source != null);
-    Contract.Invariant(cce.NonNullElements(Cases));
-    Contract.Invariant(cce.NonNullElements(MissingCases));
-  }
-
-  private Expression source;
-  private List<MatchCaseStmt> cases;
-  public readonly MatchingContext Context;
-  [FilledInDuringResolution] public readonly List<DatatypeCtor> MissingCases = new List<DatatypeCtor>();
-  public readonly bool UsesOptionalBraces;
-  public MatchStmt OrigUnresolved;  // the resolver makes this clone of the MatchStmt before it starts desugaring it
-  public MatchStmt(IToken tok, IToken endTok, Expression source, [Captured] List<MatchCaseStmt> cases, bool usesOptionalBraces, MatchingContext context = null)
-    : base(tok, endTok) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(source != null);
-    Contract.Requires(cce.NonNullElements(cases));
-    this.source = source;
-    this.cases = cases;
-    this.UsesOptionalBraces = usesOptionalBraces;
-    this.Context = context is null ? new HoleCtx() : context;
-  }
-  public MatchStmt(IToken tok, IToken endTok, Expression source, [Captured] List<MatchCaseStmt> cases, bool usesOptionalBraces, Attributes attrs, MatchingContext context = null)
-    : base(tok, endTok, attrs) {
-    Contract.Requires(tok != null);
-    Contract.Requires(endTok != null);
-    Contract.Requires(source != null);
-    Contract.Requires(cce.NonNullElements(cases));
-    this.source = source;
-    this.cases = cases;
-    this.UsesOptionalBraces = usesOptionalBraces;
-    this.Context = context is null ? new HoleCtx() : context;
-  }
-
-  public Expression Source {
-    get { return source; }
-  }
-
-  public List<MatchCaseStmt> Cases {
-    get { return cases; }
-  }
-
-  // should only be used in desugar in resolve to change the cases of the matchexpr
-  public void UpdateSource(Expression source) {
-    this.source = source;
-  }
-
-  public void UpdateCases(List<MatchCaseStmt> cases) {
-    this.cases = cases;
-  }
-
-  public override IEnumerable<Statement> SubStatements {
-    get {
-      foreach (var kase in cases) {
-        foreach (var s in kase.Body) {
-          yield return s;
-        }
-      }
-    }
-  }
-  public override IEnumerable<Expression> NonSpecificationSubExpressions {
-    get {
-      foreach (var e in base.NonSpecificationSubExpressions) { yield return e; }
-      yield return Source;
-    }
-  }
-}
-
-public class MatchCaseStmt : MatchCase {
-  private List<Statement> body;
-  public Attributes Attributes;
-  // Has the case for this constructor been generated by the resolver because the pattern was
-  // a bound variable, or was it an explicit constructor case in the source code? E.g.,
-  //
-  // var x: Option<bool>;
-  // match x
-  //   case Some(true) => ... // FromBoundVar == false
-  //   case Some(_)    => ... // FromBoundVar == false
-  //   case v          => ... // FromBoundVar == true
-  //   case _ =>       => ... // FromBoundVar == true (this case would be unreachable; added for illustration purposes)
-  //
-  // The resolved Dafny AST desugars pattern matching in a way that makes it challenging to restore the shape of the
-  // original pattern match; in particular, matching against a bound variable (or underscore) is resolved into a
-  // set of matches against all unmatched constructors. The `FromBoundVar` field provides information to code that
-  // operates on the resolved AST and that is interested in the shape of the parsed AST.
-  // This field is currently not used in the compiler but is useful for extensions and third-party compilers that
-  // use this compiler as a frontend.
-  public readonly bool FromBoundVar;
-
-  [ContractInvariantMethod]
-  void ObjectInvariant() {
-    Contract.Invariant(cce.NonNullElements(Body));
-  }
-
-  public override IEnumerable<INode> Children => body;
-
-  public MatchCaseStmt(IToken tok, DatatypeCtor ctor, bool FromBoundVar, [Captured] List<BoundVar> arguments, [Captured] List<Statement> body, Attributes attrs = null)
-    : base(tok, ctor, arguments) {
-    Contract.Requires(tok != null);
-    Contract.Requires(ctor != null);
-    Contract.Requires(cce.NonNullElements(arguments));
-    Contract.Requires(cce.NonNullElements(body));
-    this.body = body;
-    this.Attributes = attrs;
-    this.FromBoundVar = FromBoundVar;
-  }
-
-  public List<Statement> Body {
-    get { return body; }
-  }
-
-  // should only be called by resolve to reset the body of the MatchCaseExpr
-  public void UpdateBody(List<Statement> body) {
-    this.body = body;
-  }
-}
-
 /// <summary>
 /// The class represents several possible scenarios:
 /// * ...;
